Refactor schemas: move models to schemas, update imports, and create new schema files for user, receipt, and token
