+++ src/smart_cart/repositories/users.py
@@ -1,66 +1,37 @@
 from typing import Optional
 
-from sqlmodel import Field, Session, SQLModel, select
+from sqlmodel import Session, select
 
-from smart_cart.database import engine
-from smart_cart.models.user import User as UserModel
+from smart_cart.models.user import User
+from smart_cart.schemas.user import UserSchema
 from smart_cart.utils.constants import DATETIME_NOW_TIMESTAMP
-
-
-class User(SQLModel, table=True):  # type: ignore
-    user_id: str = Field(primary_key=True)
-    email: str = Field(index=True)
-    hashed_password: str = Field(min_length=8)
-    first_name: str = Field(min_length=2)
-    last_name: str = Field(min_length=2)
-    created_at: int
-    updated_at: Optional[int] = None
-    last_login: Optional[int] = None
-    is_active: bool = Field(default=True, index=True)
-    is_superuser: bool = Field(default=False)
-    is_staff: bool = Field(default=False)
-
-    @classmethod
-    def from_model(cls, item: dict) -> "User":
-        return cls(
-            user_id=item["user_id"],
-            email=item["email"],
-            hashed_password=item["hashed_password"],
-            first_name=item["first_name"],
-            last_name=item["last_name"],
-            created_at=item["created_at"],
-            updated_at=item.get("updated_at"),
-            last_login=item.get("last_login"),
-            is_active=item["is_active"],
-            is_superuser=item["is_superuser"],
-            is_staff=item["is_staff"],
-        )
+from smart_cart.utils.settings import engine
 
 
 class UserRepository:
     @staticmethod
-    def create_user(user: UserModel):
+    def create_user(user: UserSchema):
         with Session(engine) as session:
             db_user = User.from_model(user.model_dump())
             session.add(db_user)
             session.commit()
 
     @staticmethod
-    def get_user(user_id: str) -> Optional[UserModel]:
+    def get_user(user_id: str) -> Optional[UserSchema]:
         with Session(engine) as session:
             db_user = session.get(User, user_id)
-            return UserModel(**db_user.model_dump()) if db_user else None
+            return UserSchema(**db_user.model_dump()) if db_user else None
 
     @staticmethod
-    def get_user_by_email(email: str) -> Optional[UserModel]:
+    def get_user_by_email(email: str) -> Optional[UserSchema]:
         with Session(engine) as session:
             statement = select(User).where(User.email == email)
             db_user = session.exec(statement).first()
             print(f"db_user: {db_user}")
-            return UserModel(**db_user.model_dump()) if db_user else None
+            return UserSchema(**db_user.model_dump()) if db_user else None
 
     @staticmethod
-    def update_user(user: UserModel):
+    def update_user(user: UserSchema):
         with Session(engine) as session:
             db_user = session.get(User, user.user_id)
             if db_user:
@@ -77,7 +48,7 @@ class UserRepository:
                 session.commit()
 
     @staticmethod
-    def login_user(user: UserModel) -> User | UserModel:
+    def login_user(user: UserSchema) -> User | UserSchema:
         with Session(engine) as session:
             db_user = session.get(User, user.user_id)
             if db_user:
