+++ src/smart_cart/utils/auth.py
@@ -4,9 +4,9 @@ import uuid
 import bcrypt
 import jwt
 
-from smart_cart.models.token import TokenPayload
-from smart_cart.models.user import User as UserModel
 from smart_cart.repositories.users import User
+from smart_cart.schemas.token import TokenPayload
+from smart_cart.schemas.user import UserSchema
 from smart_cart.utils.constants import DATETIME_NOW, DATETIME_NOW_TIMESTAMP
 from smart_cart.utils.settings import settings
 
@@ -19,7 +19,7 @@ def verify_password(password: str, hashed_password: str) -> bool:
     return bcrypt.checkpw(password.encode("utf-8"), hashed_password.encode("utf-8"))
 
 
-def create_access_token(user: User | UserModel) -> str:
+def create_access_token(user: User | UserSchema) -> str:
     token_payload = TokenPayload(
         jti=str(uuid.uuid4()),
         sub=user.user_id,
