+++ src/tests/schemas/token_test.py
@@ -1,7 +1,7 @@
 import uuid
 
 from smart_cart.factories.token import token_payload_factory
-from smart_cart.models.token import TokenPayload
+from smart_cart.schemas.token import TokenPayload
 
 
 def test_token_payload():
